pyproject.toml
[project]
name = "documatic"
version = "0.1.0"
description = "RAG application for AppPack.io documentation with LanceDB vector storage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai",
    "pydantic-settings",
    "lancedb",
    "click",
    "httpx",
    "GitPython",
    "python-frontmatter",
    "pathlib",
    "pyarrow",
    "openai",
    "tomli-w",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
documatic = "documatic.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]
